% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame.}

\item{cl}{A true class value of your training data.}

\item{k_nn}{An integer representing the number of neighbors.}

\item{k_cv}{An integer representing the number of folds.}
}
\value{
A list with object:
\code{class}: A vector of the predicted class Ŷ i for all observation.
\code{cv_err}: A numeric with the cross-validation misclassification error.
}
\description{
This function uses k-Nearest Neighbors Cross-Validation merhosa to predict
output class using other covariates in R.
}
\examples{
data("my_penguins")
mod_penguins <- my_penguins \%>\% na.omit()
species <- as.vector(mod_penguins$species)
input_penguins <- mod_penguins \%>\% dplyr::select(bill_length_mm, bill_depth_mm,
 flipper_length_mm, body_mass_g)
my_knn_cv(input_penguins, species, 1, 5)

}
\keyword{prediction}
