% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object, similar to lm().}

\item{data}{input data frame.}
}
\value{
A table with rows for each coefficient (including the (\code{Intercept})!)
and columns for the \code{Estimate}, \verb{Std. Error}, \verb{t value}, and \verb{Pr(>|t|)}.
}
\description{
This function fits a linear model in R.
}
\examples{
data("mtcars")
my_lm(mpg ~ hp + wt, mtcars)

}
