% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Nnumber of folds.}
}
\value{
Numeric with the cross-validation error of \code{k}-fold
cross-validation.
}
\description{
This function predicts the \code{body_mass_g} of the penguins data using using
covariates \code{bill_length_mm}, \code{bill_depth_mm}, and \code{flipper_length_mm} in R.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
