% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.}

\item{mu}{Numeric indicating the null hypothesis value of the mean.}
}
\value{
List containing the numeric test statistic, the degree of freedom,
the value of the alternative, the numeric p-value.
}
\description{
This function conducts a one sample t-test.
}
\examples{
my_t.test(c(4,5,6,7,8,9), "less", 3)
my_t.test(c(11,13,17,20,33), "two.sided", 15)

}
\keyword{inference}
