% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
Vector of the predicted class for all observations.

Numeric with the cross-validation misclassification error.
}
\description{
This function predicts output class species.
}
\examples{
data(my_penguins)
my_knn_cv(na.omit(my_penguins)[, 3:6], na.omit(my_penguins)[, 1], 1, 5)
my_knn_cv(na.omit(my_penguins)[, 4:6], na.omit(my_penguins)[, 1], 2, 4)

}
\keyword{prediction}
