% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object.}

\item{data}{Input data frame.}
}
\value{
Coefficient table.
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ hp + wt, mtcars)
my_lm(mpg ~ hp + wt + drat, mtcars)

}
\keyword{prediction}
