% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.r
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation}
\usage{
my_rf_cv(k = 5)
}
\arguments{
\item{k}{numeric input, representing the number of folds used, defaults
to \code{5}}
}
\value{
A numeric with with the cross-validation error.
}
\description{
This function performs Cross Validation for Random Forest with 100 trees.
}
\examples{
## performs cross-validation with 5 folds
my_rf_cv()

## performs cross-validation with specified number of folds (e.g. 3)
k_cv <- 3
my_rf_cv(k = k_cv)
}
\keyword{prediction}
