% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.r
\name{my_t.test}
\alias{my_t.test}
\title{Student's t-test}
\usage{
my_t.test(x, alternative = "two.sided", mu = 0)
}
\arguments{
\item{x}{Numeric vector of data to be tested.}

\item{alternative}{Character string specifying the alternative hypothesis.
Accepts \code{"two.sided"}, \code{"less"}, or \code{"greater"},
defaults to \code{"two.sided"}.}

\item{mu}{Numeric input indicating the null hypothesis value of the mean,
defaults to 0.}
}
\value{
A list with the following elements: \code{test_stat} representing
the numeric test statistic, \code{df} representing the degrees
of freedom, \code{alternative} with the value of the parameter \code{alternative},
and \code{p_val} the numeric p-value.
}
\description{
This function performs a one sample Student's t-test on vector data.
}
\examples{
tdata <- rnorm(100, 5, 1)
my_t.test(tdata, mu = 5)

tdata <- rnorm(1000, 1, 1)
my_t.test(tdata, mu = 1, alternative = "greater")

}
\keyword{t.test}
