% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T.test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater". Otherwise, the
function should throw an informative error.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements:
\code{test_stat}: The numeric test statistic.
\code{df}: The degrees of freedom.
\code{alternative}: The value of \code{alternative}.
\code{p_val}: The numeric p-value.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
set.seed(100)
my_t_test(rnorm(100, mean = 0, sd = 1), "two.sided", 0)
my_t_test(rnorm(100, mean = 0, sd = 1), "less", 0)
my_t_test(rnorm(100, mean = 0, sd = 1), "greater", 0)

}
\keyword{inference}
